move slice_auto_sharded_stages to spmd & Fix an auto-sharding bug for alias
