+++ tensorflow/compiler/xla/service/spmd/BUILD
@@ -165,3 +165,16 @@ cc_library(
         "@pybind11",
     ],
 )
+
+cc_library(
+    name = "slice_auto_sharded_stages",
+    srcs = ["slice_auto_sharded_stages.cc"],
+    hdrs = ["slice_auto_sharded_stages.h"],
+    deps = [
+        "//tensorflow/compiler/xla:shape_util",
+        "//tensorflow/compiler/xla/service:hlo",
+        "//tensorflow/compiler/xla/service:hlo_pass",
+        "//tensorflow/compiler/xla/service:pass_context",
+        "@pybind11",
+    ],
+)
