+++ tensorflow/compiler/xla/service/spmd/slice_auto_sharded_stages.h
@@ -1,10 +1,10 @@
-#ifndef TENSORFLOW_COMPILER_XLA_SERVICE_GPU_SLICE_AUTO_SHARDED_STAGES
-#define TENSORFLOW_COMPILER_XLA_SERVICE_GPU_SLICE_AUTO_SHARDED_STAGES
+#ifndef TENSORFLOW_COMPILER_XLA_SERVICE_SPMD_SLICE_AUTO_SHARDED_STAGES
+#define TENSORFLOW_COMPILER_XLA_SERVICE_SPMD_SLICE_AUTO_SHARDED_STAGES
 
 #include "tensorflow/compiler/xla/service/hlo_pass_interface.h"
 
 namespace xla {
-namespace gpu {
+namespace spmd {
 
 class SliceAutoShardedStages : public HloModulePass {
  public:
@@ -17,7 +17,7 @@ class SliceAutoShardedStages : public HloModulePass {
   StatusOr<bool> Run(HloModule* module) override;
 };
 
-}  // namespace gpu
+}  // namespace spmd
 }  // namespace xla
 
-#endif  // TENSORFLOW_COMPILER_XLA_SERVICE_GPU_SLICE_AUTO_SHARDED_STAGES
+#endif  // TENSORFLOW_COMPILER_XLA_SERVICE_SPMD_SLICE_AUTO_SHARDED_STAGES
